fix issues with evidence codes * multiple code sames ref * tooltip flickering in and out with imprecise placement of cursor * layout for multiple codes
